Query cache: Enable squash of partial result blocks
Compression of query cache entries (TBD in a future commit) will be more
effective if the input blocks are not too tiny. Setting
"query_cache_squash_partial_query_results" allows to concatenate the
input blocks to a single large block. The setting is disabled for now,
because the performance implications are not clear.

A follow-up idea (added as TODO comment) is to squash only down to
max_block_size. Yet most partial result chunks should not surpass this
size already now.